+++ test/testUpdateMKD.jl
@@ -14,7 +14,7 @@ using LieGroups
     m1 = manikde!(M, pts)
 
     pts = [randn(2) for _ = 1:100]
-    m2 = manikde!(M, pts; infoPerCoord = [0.3; 0.4])
+    m2 = manikde!(M, pts; observability = [0.3; 0.4])
 
     @test 0 < mmd(m1, m2)
 
@@ -22,7 +22,7 @@ using LieGroups
 
     # @test mmd(m1, m2) < 1e-6
 
-    # @test isapprox(m1.infoPerCoord, m2.infoPerCoord)
+    # @test isapprox(m1.observability, m2.observability)
 
     ##
 end